family form refactor: separate section components, consistent layout, save state on change
