+++ app/src/components/form/FormProgressBar.tsx
@@ -1,10 +1,11 @@
-import { useLocation, useNavigate } from "@tanstack/react-router";
+import { Link, useLocation, useParams } from "@tanstack/react-router";
 import {
   DocumentCheckIcon,
   GiftIcon,
   UserIcon,
   UsersIcon,
 } from "@heroicons/react/24/solid";
+import type { ReactNode } from "react";
 import type { FamilyFormState } from "@/components/providers/FormProvider";
 import { useFormContext } from "@/components/providers/FormProvider";
 import { SECTION_SCHEMAS } from "@/lib/formSchemas";
@@ -12,65 +13,106 @@ import { SECTION_SCHEMAS } from "@/lib/formSchemas";
 type FormStep = {
   id: string;
   label: string;
-  path: string;
   icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
   sectionKey: keyof FamilyFormState;
-  isReviewPage?: boolean; // Flag for pages without their own form data
 };
 
 const FORM_STEPS: Array<FormStep> = [
   {
     id: "general",
     label: "General",
-    path: "/family/form/general-info",
     icon: UserIcon,
     sectionKey: "generalInfo",
   },
   {
     id: "children",
     label: "Child Info",
-    path: "/family/form/children",
     icon: UsersIcon,
     sectionKey: "children",
   },
   {
     id: "gifts",
     label: "Gifts!",
-    path: "/family/form/gift-details",
     icon: GiftIcon,
     sectionKey: "gifts",
   },
   {
     id: "review",
     label: "Review",
-    path: "/family/form/review",
     icon: DocumentCheckIcon,
-    sectionKey: "generalInfo", // Review doesn't have its own section
-    isReviewPage: true, // Flag to identify review page
+    sectionKey: "generalInfo",
   },
 ];
 
-type StepState = "current" | "complete" | "error" | "incomplete";
+type StepLinkProps = {
+  stepId: string;
+  driveId: string;
+  className: string;
+  children: ReactNode;
+};
 
-interface FormProgressBarProps {
-  onNavigate?: (targetPath: string) => void | Promise<void>;
+function StepLink({ stepId, driveId, className, children }: StepLinkProps) {
+  const params = { driveId };
+  switch (stepId) {
+    case "general":
+      return (
+        <Link
+          to="/family/drive/$driveId/form/general-info"
+          params={params}
+          className={className}
+        >
+          {children}
+        </Link>
+      );
+    case "children":
+      return (
+        <Link
+          to="/family/drive/$driveId/form/children"
+          params={params}
+          className={className}
+        >
+          {children}
+        </Link>
+      );
+    case "gifts":
+      return (
+        <Link
+          to="/family/drive/$driveId/form/gift-details"
+          params={params}
+          className={className}
+        >
+          {children}
+        </Link>
+      );
+    case "review":
+      return (
+        <Link
+          to="/family/drive/$driveId/form/review"
+          params={params}
+          className={className}
+        >
+          {children}
+        </Link>
+      );
+    default:
+      return <div className={className}>{children}</div>;
+  }
 }
 
-export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
+type StepState = "current" | "complete" | "error" | "incomplete";
+
+export function FormProgressBar() {
   const { formState } = useFormContext();
   const location = useLocation();
-  const navigate = useNavigate();
+  const { driveId } = useParams({ strict: false });
   const currentPath = location.pathname;
 
   const getStepState = (step: FormStep, index: number): StepState => {
-    // Check if this is the current step
-    if (currentPath === step.path) return "current";
+    if (currentPath.includes(`/form/${getPathSegment(step.id)}`)) return "current";
     return getUnderlyingState(step, index);
   };
 
-  // Helper to get the underlying state (ignoring current status)
   const getUnderlyingState = (step: FormStep, _: number): StepState => {
-    // Special handling for review step
     if (step.id === "review") {
       const allPreviousComplete = FORM_STEPS.slice(0, -1).every((s) => {
         const data = formState[s.sectionKey];
@@ -81,13 +123,9 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
       return allPreviousComplete ? "complete" : "incomplete";
     }
 
-    // Check if step has data
     const stepData = formState[step.sectionKey];
-    if (!stepData) {
-      return "incomplete";
-    }
+    if (!stepData) return "incomplete";
 
-    // Validate with schema if available
     const schema = SECTION_SCHEMAS[step.sectionKey];
     return schema.safeParse(stepData).success ? "complete" : "error";
   };
@@ -104,24 +142,10 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
     return index < firstUnvisitedIndex;
   };
 
-  const handleStepClick = async (step: FormStep, index: number) => {
-    if (!isStepClickable(step, index)) {
-      return;
-    }
-
-    // This allows the parent form to save current state before navigating
-    if (onNavigate) {
-      await onNavigate(step.path);
-    } else {
-      // Default navigation
-      navigate({ to: step.path as any });
-    }
-  };
-
   const getStepStyles = (state: StepState, underlyingState: StepState) => {
     switch (state) {
       case "current": {
-        let fillColor = "bg-gray-300"; // default
+        let fillColor = "bg-gray-300";
         let iconColor = "text-gray-500";
 
         if (underlyingState === "complete") {
@@ -134,10 +158,10 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
 
         return {
           iconBg: fillColor,
-          iconBorder: "border-[#F4D03F] border-[3px]", // Gold border
+          iconBorder: "border-[#F4D03F] border-[3px]",
           iconColor: iconColor,
           labelColor: "text-gray-900",
-          underline: "border-b-2 border-[#F4D03F]", // Gold underline
+          underline: "border-b-2 border-[#F4D03F]",
         };
       }
       case "complete":
@@ -180,8 +204,6 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
             <div className="absolute inset-0 flex">
               {FORM_STEPS.slice(0, -1).map((step, index) => {
                 const currentState = getStepState(step, index);
-
-                // Line is colored if current step is complete or current
                 const shouldColor =
                   currentState === "complete" || currentState === "current";
                 const lineColor =
@@ -209,28 +231,21 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
             const Icon = step.icon;
             const clickable = isStepClickable(step, index);
 
-            return (
-              <div
-                key={step.id}
-                className="flex flex-col items-center relative z-10"
-                onClick={() => handleStepClick(step, index)}
-              >
-                {/* Icon circle */}
+            const stepContent = (
+              <>
                 <div
                   className={`
-                    w-[44px] h-[44px] rounded-full 
+                    w-[44px] h-[44px] rounded-full
                     ${styles.iconBorder}
                     ${styles.iconBg}
                     flex items-center justify-center
                     transition-all duration-300
                     shadow-sm
-                    ${clickable ? "cursor-pointer hover:scale-110 hover:shadow-md" : "cursor-default"}
+                    ${clickable ? "hover:scale-110 hover:shadow-md" : ""}
                   `}
                 >
                   <Icon className={`w-5 h-5 ${styles.iconColor}`} />
                 </div>
-
-                {/* Label with conditional underline */}
                 <div className="mt-2.5 text-center">
                   <span
                     className={`
@@ -239,12 +254,29 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
                       inline-block pb-0.5
                       transition-colors duration-300
                       whitespace-nowrap
-                      ${clickable ? "cursor-pointer" : "cursor-default"}
                     `}
                   >
                     {step.label}
                   </span>
                 </div>
+              </>
+            );
+
+            return clickable ? (
+              <StepLink
+                key={step.id}
+                stepId={step.id}
+                driveId={driveId ?? ""}
+                className="flex flex-col items-center relative z-10 cursor-pointer"
+              >
+                {stepContent}
+              </StepLink>
+            ) : (
+              <div
+                key={step.id}
+                className="flex flex-col items-center relative z-10 cursor-default"
+              >
+                {stepContent}
               </div>
             );
           })}
@@ -252,4 +284,14 @@ export function FormProgressBar({ onNavigate }: FormProgressBarProps) {
       </div>
     </div>
   );
+}
+
+function getPathSegment(stepId: string): string {
+  switch (stepId) {
+    case "general": return "general-info";
+    case "children": return "children";
+    case "gifts": return "gift-details";
+    case "review": return "review";
+    default: return "";
+  }
 }
